+++ waspc/cli/src/Wasp/Cli/Command/Start.hs
@@ -8,17 +8,20 @@ import Control.Concurrent.MVar (MVar, newMVar, tryTakeMVar)
 import Control.Monad.Except (throwError)
 import Control.Monad.IO.Class (liftIO)
 import StrongPath ((</>))
+import Wasp.Cli.AppComponents (makeDevRunConfigs)
 import Wasp.Cli.Command (Command, CommandError (..), require)
 import Wasp.Cli.Command.Compile (compile, printWarningsAndErrorsIfAny)
+import Wasp.Cli.Command.LockedProject (withLockedProject)
 import Wasp.Cli.Command.Message (cliSendMessageC)
 import Wasp.Cli.Command.News (fetchAndListMustSeeNewsIfDue)
 import Wasp.Cli.Command.Require.DbConnectionEstablished (DbConnectionEstablished (DbConnectionEstablished))
 import Wasp.Cli.Command.Require.InWaspProject (InWaspProject (InWaspProject))
 import Wasp.Cli.Command.Watch (watch)
-import Wasp.Cli.Services (devEnvVars, devPorts, devUrls)
 import Wasp.Cli.Util.EnvVarInputs (resolveEnvVarInputs)
 import qualified Wasp.Cli.Util.EnvVarInputs as EnvVarInputs
 import qualified Wasp.Generator
+import qualified Wasp.Generator.Client as Client
+import qualified Wasp.Generator.Server as Server
 import qualified Wasp.Message as Msg
 import Wasp.Project (CompileError, CompileWarning)
 import Wasp.Project.Common (generatedAppDirInWaspProjectDir)
@@ -27,7 +30,7 @@ import qualified Wasp.Project.Env as Env
 -- | Does initial compile of wasp code and then runs the generated project.
 -- It also listens for any file changes and recompiles and restarts generated project accordingly.
 start :: Command ()
-start = do
+start = withLockedProject $ do
   -- We check for the news only in `wasp start`, and only periodically,
   -- to avoid being too aggressive. Specifically:
   --   - We don't run it in other `wasp` commands because we don't want to
@@ -46,12 +49,12 @@ start = do
 
   (warnings, appSpec) <- compile
 
-  let waspEnvVars = devEnvVars devPorts (devUrls appSpec devPorts)
-      envVarInputs = getEnvVarInputs <$> Env.dotEnvFiles
+  let (client, server) = makeDevRunConfigs appSpec Client.defaultPort Server.defaultPort
 
-  envVars <-
-    sequence $
-      resolveEnvVarInputs waspProjectDir <$> waspEnvVars <*> envVarInputs
+  clientEnvVars <-
+    resolveEnvVarInputs waspProjectDir (Client.devEnvVars client) (devEnvVarInputs Env.dotEnvClient)
+  serverEnvVars <-
+    resolveEnvVarInputs waspProjectDir (Server.devEnvVars server) (devEnvVarInputs Env.dotEnvServer)
 
   DbConnectionEstablished <- require
 
@@ -67,7 +70,7 @@ start = do
     ongoingCompilationResultMVar <- newMVar (warnings, [])
     let watchWaspProjectSource = watch waspProjectDir outDir ongoingCompilationResultMVar
     let startGeneratedWebApp =
-          Wasp.Generator.start envVars waspProjectDir outDir (onJobsQuietDown ongoingCompilationResultMVar)
+          Wasp.Generator.start clientEnvVars serverEnvVars waspProjectDir outDir (onJobsQuietDown ongoingCompilationResultMVar)
     -- In parallel:
     -- 1. watch for any changes in the Wasp project, be it users wasp code or users JS/HTML/...
     --    code. On any change, Wasp is recompiled (and generated app is re-generated).
@@ -82,7 +85,7 @@ start = do
       Left startError -> throwError $ CommandError "Start failed" startError
       Right () -> error "This should never happen, start should never end but it did."
   where
-    getEnvVarInputs file = [EnvVarInputs.FromProjectFile file, EnvVarInputs.Inherit]
+    devEnvVarInputs dotEnvFile = [EnvVarInputs.FromProjectFile dotEnvFile, EnvVarInputs.Inherit]
 
     onJobsQuietDown :: MVar ([CompileWarning], [CompileError]) -> IO ()
     onJobsQuietDown ongoingCompilationResultMVar = do